Move tests for Event class to it's own file
